pyproject.toml
[project]
name = "testpydantic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "devtools>=0.12.2",
    "pydantic-ai>=0.1.6",
    "python-dotenv>=1.1.0",
]
